Difficulty Manager deleted!
Improvements:

+ Difficulty Manager only calculate the GridSize. This can be done in GridManager which is specifically responsible for Grid System.
