+++ .idea/.idea.ExplodingStrides/.idea/encodings.xml
@@ -2,5 +2,6 @@
 <project version="4">
   <component name="Encoding" addBOMForNewFiles="with BOM under Windows, with no BOM otherwise">
     <file url="file://$PROJECT_DIR$/Assets/Scripts/GridManager.cs" charset="windows-1254" />
+    <file url="file://$PROJECT_DIR$/Assets/Scripts/TileController.cs" charset="windows-1254" />
   </component>
 </project>
