--- Assets/Scripts/Difficulty Manager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4d6ec43c612d46840a2001052c13f450
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
