+++ Assets/Scripts/GridManager.cs
@@ -1,9 +1,15 @@
 using UnityEngine;
 using System.Collections.Generic;
 using TMPro;
+using UnityEngine.SceneManagement;
+
 public class GridManager : MonoBehaviour
 {
     public static GridManager Instance { get; private set; }
+    
+    [SerializeField]
+    private Transform _gridStartPositionTransform;
+    
     public List<Vector2Int> MinePositions { get => minePositions; set => minePositions = value; }
 
     public static int gridSizeX;
@@ -28,15 +34,25 @@ public class GridManager : MonoBehaviour
 
     private void Start()
     {
-        CreateGrid();
-        Debug.Log("<color=magenta> Ben caliscam simdi</color>");
-        GridTransformController(transform);
-       // Vector3 newPosition = new Vector3(-4.63f, 0f, -8.18f);
-        //transform.position = newPosition;
+        UIController.Instance.GetDifficultySlider().onValueChanged.AddListener(SetGridSize);
 
-        PlaceMines();
-        PrintMinePositions();
-        TileController.PrintNeighbourCoordinates();
+        if (SceneManager.GetActiveScene().buildIndex == 1)
+        {
+            CreateGrid();
+            Debug.Log("<color=magenta> Ben caliscam simdi</color>");
+            GridTransformController(_gridStartPositionTransform);
+            // Vector3 newPosition = new Vector3(-4.63f, 0f, -8.18f);
+            //transform.position = newPosition;
+
+            PlaceMines();
+            PrintMinePositions();
+            TileController.PrintNeighbourCoordinates();
+        }
+    }
+    
+    private void OnDisable()
+    {
+        UIController.Instance.GetDifficultySlider().onValueChanged.RemoveListener(SetGridSize);
     }
 
     private void CreateGrid()
@@ -50,18 +66,46 @@ public class GridManager : MonoBehaviour
             for (int x = 0; x < gridSizeX; x++)
             {
                 Vector3 position = new Vector3(x * _cellSize, 0f, y * _cellSize);
-                Tile newTile = Instantiate(tilePrefab);
+                Tile newTile = Instantiate(tilePrefab, _gridStartPositionTransform.position,
+                    Quaternion.Euler(Vector3.right * -90f), _gridStartPositionTransform);
                 newTile.GetCoordinates(x,y);
                 _grid[x, y] = newTile;
 
-                newTile.transform.parent = transform;
                 newTile.transform.localPosition = position;
                 //_grid[x, y].tileController.UpdateMineCount();
             }
         }
 
-        Vector3 centerOffset = new Vector3(gridSizeX * _cellSize * 0.5f, gridSizeY * _cellSize * 0.5f, 0);
-        transform.position = -centerOffset;
+        /*Vector3 centerOffset = new Vector3(gridSizeX * _cellSize * 0.5f, gridSizeY * _cellSize * 0.5f, 0);
+        _gridStartPositionTransform.position = -centerOffset;*/
+    }
+
+    private void SetGridSize(float sliderValue)
+    {
+        int selectedDifficulty = Mathf.RoundToInt((int) sliderValue);
+
+        int gridSizeX = 5; // Default
+        int gridSizeY = 5; 
+
+        if (selectedDifficulty == 1) // Easy
+        {
+            gridSizeX = 5;
+            gridSizeY = 5;
+        }
+        else if (selectedDifficulty == 2) // Medium
+        {
+            gridSizeX = 7;
+            gridSizeY = 7;
+        }
+        else if (selectedDifficulty == 3) // Hard
+        {
+            gridSizeX = 10;
+            gridSizeY = 10;
+        }
+
+        PlayerPrefs.SetInt("Grid Size X", gridSizeX);
+        PlayerPrefs.SetInt("Grid Size Y", gridSizeY);
+        print(PlayerPrefs.GetInt("Grid Size X"));
     }
 
     private void PlaceMines()
@@ -114,7 +158,7 @@ public class GridManager : MonoBehaviour
         return clickedTiles.Contains(tileCoordinates);
     }
 
-    public void GridTransformController(Transform GridManagerTransform) // yer ayarlama for grid clones
+    private void GridTransformController(Transform GridManagerTransform) // yer ayarlama for grid clones
     {
         if (gridSizeX == 5 && gridSizeY == 5)
         {
